move CredentialSchema into own file
